requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-debug-toolbar==3.2.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
